LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-canvas-element/type.prototype-expected.txt

type.prototype
window.HTMLCanvasElement has prototype, which is { ReadOnly, DontDelete }. prototype has getContext, which is not
Defined in "Web IDL" (draft)
Actual output:

PASS window.HTMLCanvasElement has prototype, which is { ReadOnly, DontDelete }. prototype has getContext, which is not
